refactor banner constants
