+++ main.go
@@ -253,8 +253,8 @@ func main() {
 		os.Exit(1)
 	}
 
-	sess.Out.Info("%s\n\n", core.ASCIIBanner)
-	sess.Out.Important("%s v%s started at %s\n", core.Name, core.Version, sess.Stats.StartedAt.Format(time.RFC3339))
+	sess.Out.Info("%s\n\n", common.ASCIIBanner)
+	sess.Out.Important("%s v%s started at %s\n", common.Name, common.Version, sess.Stats.StartedAt.Format(time.RFC3339))
 	sess.Out.Important("Loaded %d signatures\n", len(core.Signatures))
 	sess.Out.Important("Web interface available at http://%s:%d\n", *sess.Options.BindAddress, *sess.Options.Port)
